+++ index.html
@@ -17,7 +17,7 @@
                         <h1>What should <br>I read next?</h1>
                         <br>
                         <br>
-                        <p>If you're on a hunt for your next great read, we can help. something something something... Books that match your criteria will be added to the "Books You Are Looking For" section.</p>
+                        <p>If you're on a hunt for your next great read, you're in the right place. This page allows you to search for books based on varius keywords, like title, author, category, etc. Simply type something into the 'Search For Books' filed and press the 'Find Books' button. Books that match your criteria will be added to the "Books You Are Looking For" section.</p>
                         <br>
                         <br>
                         <a href="#book-finder" class="btn btn__header">Search For Books</a>
@@ -39,81 +39,14 @@
         </section>
         <section id="books-grid">
             <h3>Books You Are Looking For</h3>
-            <div class="container__books-grid">
-             
-                <!-- <div class="books-grid__card">
-                    <div class="books-grid__img"> -->
-                        <!-- <img src="" alt=""> -->
-                    <!-- </div>
-                    <div class="books-grid__decs"> -->
-                        <!-- <h4></h4>
-                        <p class ="author"></p>
-                        <p></p> -->
-                    <!-- </div>
-                    <div class="books-grid__card--overlay">
-                        <a href="">Find out more</a>
-                    </div>
-                </div>  -->
-                <!-- <div class="books-grid__card">
-                    <div class="books-grid__img">
-                        <img src="" alt="">
-                    </div>
-                    <div class="books-grid__decs">
-                        <h4>Book Title Goes Here</h4>
-                        <p class ="author">AUTHOR NAME</p>
-                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris neque
-                            tellus, feugiat at ultricies at, vulputate at augue. Sed in leo congue,
-                            vestibulum libero laoreet.</p>
-                    </div>
-                    <div class="books-grid__card--overlay">
-                        <a href="">Find out more</a>
-                    </div>
-                </div> -->
-               <!-- <div class="books-grid__card">
-                    <div class="books-grid__img">
-                        <img src="" alt="">
-                    </div>
-                    <div class="books-grid__decs">
-                        <h4>Book Title Goes Here</h4>
-                        <p class ="author">AUTHOR NAME</p>
-                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris neque
-                            tellus, feugiat at ultricies at, vulputate at augue. Sed in leo congue,
-                            vestibulum libero laoreet.</p>
-                    </div>
-                    <div class="books-grid__card--overlay">
-                        <a href="">Find out more</a>
-                    </div>
-                </div> -->
-                <!-- <div class="books-grid__card">
-                    <div class="books-grid__img">
-                        <img src="" alt="">
-                    </div>
-                    <div class="books-grid__decs">
-                        <h4>Book Title Goes Here</h4>
-                        <p class ="author">AUTHOR NAME</p>
-                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris neque
-                            tellus, feugiat at ultricies at, vulputate at augue. Sed in leo congue,
-                            vestibulum libero laoreet.</p>
-                    </div>
-                </div> -->
+            <div class="container__books-grid">         
             </div>
-            
-
-<!-- <div id="myModal" class="modal">
-
-
-    <div class="modal-content">
-      <span class="close">&times;</span>
-      <p>Some text in the Modal..</p>
-    </div>
-  
-  </div> -->
   
         </section>
     </main>
 <footer>
     <p class="text text--reg text--light-center"> &copy Copyright Martyna Krol <script>document.write(new Date().getFullYear());</script></p>
 </footer>
-    <script src="./../new.js"></script>
+    <script type = "module" src="./../main.js"></script>
 </body>
 </html>
