+++ modules/books-grid.scss
@@ -43,27 +43,12 @@ h4 {
 	flex-direction: column;
 	align-items: center;
 	text-align: left;
-	cursor: pointer;
+	// cursor: pointer;
 }
 
 .container__books-grid > div > * {
 	margin: 10px;
 }
-
-// .books-grid__card {
-// 	position: relative;
-// 	display: flex;
-// 	flex-direction: column;
-// 	border: 1px solid black;
-// 	padding: 15px;
-// 	@include mixins.phone {
-// 		flex-direction: column;
-// 	}
-// 	&:hover {
-// 		border: none;
-// 		transition: 0.5s ease;
-// 	}
-// }
 
 .container__books-grid > div > img {
 	height: 200px;
@@ -89,35 +74,10 @@ h4 {
 	}
 }
 
-// .books-grid__decs p {
-// 	font-size: 12px;
-// }
-
 .author {
 	font-size: 16px;
 	text-transform: uppercase;
 }
-// 	padding-bottom: 15px;
-// 	@include mixins.phone {
-// 		padding-bottom: 15px;
-// 	}
-// }
-
-// .books-grid__card--overlay {
-// 	position: absolute;
-// 	top: 0;
-// 	bottom: 0;
-// 	left: 0;
-// 	right: 0;
-// 	height: 100%;
-// 	width: 100%;
-// 	opacity: 0;
-// 	transition: 0.5s ease;
-// 	background-color: palette.$peach;
-// 	display: flex;
-// 	justify-content: center;
-// 	align-items: center;
-// }
 
 .books-grid__card:hover .books-grid__card--overlay {
 	opacity: 0.95;
@@ -135,7 +95,7 @@ a {
 
 /* The Modal (background) */
 .modal {
-	display: none; /* Hidden by default */
+	// display: none; /* Hidden by default */
 	position: fixed; /* Stay in place */
 	z-index: 1; /* Sit on top */
 	padding-top: 100px; /* Location of the box */
@@ -145,7 +105,8 @@ a {
 	height: 100vh; /* Full height */
 	overflow: auto; /* Enable scroll if needed */
 	background-color: rgb(0, 0, 0); /* Fallback color */
-	background-color: rgba(0, 0, 0, 0.1); /* Black w/ opacity */
+	background-color: rgba(0, 0, 0, 0.8); /* Black w/ opacity */
+	display: none;
 }
 
 /* Modal Content */
@@ -156,9 +117,25 @@ a {
 	border: 1px solid black;
 	width: 40%;
 	height: auto;
+	@include mixins.landscape-tablet-and-smaller {
+		width: 60%;
+	}
 }
 
 .modal-content > p {
 	padding: 20px;
 	text-align: left;
+}
+
+.modal-content > h4 {
+	padding-bottom: 20px;
+}
+
+.close {
+	background-color: white;
+	border: none;
+	@extend %para;
+	margin-bottom: 20px;
+	font-size: 30px;
+	cursor: pointer;
 }
