+++ modules/header.scss
@@ -103,10 +103,6 @@ p {
 	}
 }
 
-.btn-again {
-	margin: 20px;
-}
-
 #btn {
 	&.invisible {
 		display: none;
