+++ modules/search-form.scss
@@ -54,8 +54,6 @@ form {
 	padding: 45px;
 }
 
-.search-form__btn {
-	&.invisible {
-		display: none;
-	}
+.error-para {
+	padding-bottom: 20px;
 }
